+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Generate `Default_*` methods that construct Thrift structs with defined 
+  default values pre-populated.  
+
 ### Changed
 - gen: Redefine Options.Plugin as a struct usable outside go.uber.org/thriftrw.